refactor: re-add support compiling with QT < 5.14
The refactor to remove deprecated symbols also broke compatibility with versions of QT prior to QT5.14.  This commit adds back support for earlier versions of QT. Mostly this is via the new "qtsupport.h" header file, but for supporting the changed `QSet` constructor we had to put the changes into each source file. This is, of course, not optimal.