build: update `cl-format.sh` to force v13 of clang-format
The script was designed to find the latest version of clang-format, however, this project currently specifically requires the use of v13. We now only check for v13 of clang-format.

Secondly, for cmake-format the script also requires a specific version. Instead of installing that unmanaged version globally we now install it inside a python venv.